Create K-sorted_arrays.cpp
You are working on a project that involves merging K-sorted arrays into a single sorted array. Each array represents the scores of students from different schools in a programming competition. 



Your task is to develop a program that takes input from the user for K-sorted arrays and merges them into a single sorted array, which will be used for further analysis.



Note: This is a sample question asked in TCS recruitment.

Input format :
The first line of input consists of the number of sorted arrays K to merge.

The next line of input consists of n - the size of the first sorted array, followed by n elements.

The same is repeated for K arrays.

Output format :
The output prints the merged arrays in descending order, separated by space.

Sample test cases :
Input 1 :
3
3
1 5 7
2
2 4
3
6 8 9
Output 1 :
9 8 7 6 5 4 2 1 
Input 2 :
3
5
2 4 6 8 10
3
1 3 5
4
7 9 11 13
Output 2 :
13 11 10 9 8 7 6 5 4 3 2 1