Create retail_company_ merging_array.cpp
You are working as a software developer for an online retail company. The company has tasked you with creating a program to merge two arrays that contain product IDs. The first array, stock, represents the product IDs of items currently in stock. 



The second array, cart, represents the product IDs of items in a customer's shopping cart. Your goal is to merge these two arrays into a single, sorted list of unique product IDs.

Input format :
The first line contains the size of the first array.

The second line contains the elements of the first array, in ascending order.

The third line contains the size of the second array.

The fourth line contains the elements of the second array, in ascending order.

Output format :
The output displays the merged array, which is the ascending order of the given array.

Code constraints :
The maximum size of each array is limited to 100 (const int MAX_SIZE = 100).

The size of each array should be a positive integer.

The elements of each array can be any integer.

The code assumes that the user will input valid integers as elements of the arrays.

Sample test cases :
Input 1 :
5
-1 2 3 4 5
4
-6 7 8 9
Output 1 :
-6 -1 2 3 4 5 7 8 9 
Input 2 :
6
2 4 6 8 10 12
5
2 4 9 10 11
Output 2 :
2 4 6 8 9 10 11 12