Create police_man_search_record.cpp
A policeman wants to search for a particular record in a file containing new records. He needs to determine whether the record he is looking for exists in the file or not. Help the policeman by writing a program that takes the number of records, the record values, and the item to be searched as input using dynamic memory allocation. 



Implementing the program using linear search to display whether the item is found in the records or not.

Input format :
The first line contains an integer n representing the number of elements in the array.

The second line contains n space-separated integers representing the elements of the array.

The third line contains an integer item representing the item to be searched in the array.

Output format :
The output is a single line that either says "Item found" if the item is found in the array or "Item not found" if the item is not found.

Code constraints :
1<=n<1000

Sample test cases :
Input 1 :
5
4 2 3 5 1
2
Output 1 :
Item found
Input 2 :
5
6 7 8 9 2
4
Output 2 :
Item not found