Create  findMedian_using_binary_search.cpp
You are working on a data analysis project that involves analyzing a two-dimensional matrix representing data measurements. Each row and column of the matrix corresponds to a specific data point, and the matrix is sorted in ascending order both row-wise and column-wise.



As a part of the data analysis process, you need to implement a function that finds the median value of the data represented by the sorted matrix using binary search.



Write a function findMedian() that takes in the sorted matrix of integers, the number of rows, and the number of columns as inputs. The function should use binary search to find the median value of the data in the matrix.



Assume that the number of elements in the matrix is always odd, so there is a definite median value. The function should return the median value found.



Note: This is a sample question asked in Capgemini recruitment.

Input format :
The first line of input consists of the number of rows m and columns n, separated by space.

The following lines of input consist of m*n elements.

Output format :
The output prints the median value in the given array.

Sample test cases :
Input 1 :
3 3
1 2 3
4 5 6
7 8 9
Output 1 :
5
Input 2 :
5 5
10 11 12 13 15
21 22 26 29 31
32 34 36 38 40
41 42 43 44 45
56 57 58 59 60
Output 2 :
36