Create findPeakElement_using_binary_search.cpp
You are a mountaineer exploring a range of mountains that are known for their unique structure. The mountains in this range have a distinct pattern where they increase in height up to a certain point and then start decreasing. Your goal is to find the peak of the highest mountain in this range.



To assist you in this task, you decide to develop a program that uses binary search to efficiently find the peak element in a given mountain array.



Write a function findPeakElement() that takes in a mountain array of integers as input. The mountain array is guaranteed to have a peak element, which is the highest value in the array. The function should use binary search to find and return the peak element.



Note: This is a sample question asked in Infosys recruitment.

Input format :
The first line of input consists of the number of elements in the array.

The second line of input consists of the mountain array elements, separated by space.

Output format :
The output prints the peak element in the given array.

Sample test cases :
Input 1 :
8
0 2 3 4 5 2 1 0
Output 1 :
5
Input 2 :
5
1 2 3 0 -6
Output 2 :
3