Create  findSmallestMissing_using_binary_search.cpp
You are working on a task management application that allows users to prioritize their tasks by assigning them positive integer values. The application maintains a sorted list of task priorities in ascending order to help users manage their workload effectively.



As a part of the task management feature, you need to implement a function that finds the smallest missing positive integer greater than a given value x in the sorted array of task priorities using binary search. This will allow users to determine the next available priority value for assigning new tasks.



Write a function findSmallestMissing() that takes in the sorted array of task priorities, the size of the array, and the given value x as inputs. The function should use binary search to find the smallest missing positive integer greater than x in the array.



Note: This is a sample question asked in TCS recruitment.

Input format :
The first line of input consists of the number of elements in the array.

The second line of input consists of array elements, separated by space.

The third line of input consists of the value of x.

Output format :
The output prints the smallest missing positive integer greater than the given value x.

Sample test cases :
Input 1 :
6
12 24 26 37 41 53
24
Output 1 :
The smallest missing positive integer greater than 24 is 25.
Input 2 :
6
12 24 24 37 41 53
52
Output 2 :
The smallest missing positive integer greater than 52 is 54.