Create kth_smallest_ticket_price.cpp
You are working on a ticketing system for a popular event. As part of the system, you need to implement a feature that finds the kth smallest ticket price among all the available tickets.



Write a program that takes user input for the number of tickets and their respective prices. The program should use the selection sort algorithm to sort the ticket prices in non-decreasing order. 



After sorting, it should display the sorted order of the ticket prices. Additionally, the program should find and display the kth smallest ticket price from the sorted list.



This feature will help event organizers and attendees identify the kth cheapest ticket price, which can be useful for various purposes such as determining different pricing tiers or identifying affordable ticket options. The program provides an efficient way to sort and find the kth smallest ticket price within the ticketing system, enhancing the overall ticket-purchasing experience for event participants.



Note: This is a sample question asked in Capgemini recruitment.

Input format :
The first line of input consists of the number of elements n in the array.

The second line of input consists of n array elements, separated by space.

The third line of input consists of the value of k.

Output format :
The first line of output prints the sorted order.

The second line of output prints the kth smallest element in the given array.



Refer to the sample output for formatting specifications.

Code constraints :
n>0

k<n

Sample test cases :
Input 1 :
5
1 4 7 8 5
3
Output 1 :
Sorted order: 1 4 5 7 8 
The kth smallest element is: 5
Input 2 :
9
7 9 6 8 4 5 3 2 1
6
Output 2 :
Sorted order: 1 2 3 4 5 6 7 8 9 
The kth smallest element is: 6