Create linked_list_inserting_at_front.cpp
Alice is learning about linked lists and wants to practice creating one by inserting elements at the front. She decides to write a program that allows her to do so. 



The program will prompt Alice to enter the number of elements she wants to insert and then ask her to input the values of those elements. After receiving the inputs, the program will insert the elements at the front of the linked list and display the final linked list to Alice.



Alice is excited to see how her program will work and hopes it will help solidify her understanding of linked lists. She prepares herself to input the number of elements and their respective values, eagerly anticipating the final result.



Note: This is a sample question asked in a HCL interview.

Input format :
The first line of input consists of an integer n, representing the number of elements to be inserted.

The second line of input consists of n space-separated integers, representing the elements to be inserted in the linked list.

The third line of input consists of an integer representing the value to be inserted at the front of the linked list.

Output format :
The first line of output should display the initial linked list elements, separated by space.

The second line of output should display the linked list elements, after inserting the given element at the front, separated by space.



Refer to the sample output for formatting specifications.

Code constraints :
The number of elements to be inserted (n) is a positive integer.

The values of the elements can be positive or negative integers.

Sample test cases :
Input 1 :
5
6 5 4 3 2
1
Output 1 :
Created Linked list: 2 3 4 5 6 
Final List: 1 2 3 4 5 6 
Input 2 :
4
-10 20 -30 40
5
Output 2 :
Created Linked list: 40 -30 20 -10 
Final List: 5 40 -30 20 -10