Create adds_1_linked_list.cpp
You are working on a program for an inventory management system in a retail store. The store uses barcodes to label and track its products. Each barcode is represented as a linked list, where each digit of the barcode is stored in a separate node. Your task is to write a function that adds 1 to the barcode value and updates the linked list accordingly.



For example, 1999 is represented as (1-> 9-> 9 -> 9), and adding 1 to it should change it to (2->0->0->0) 



Note: This is a sample question asked in the Flipkart interview.

Input format :
The first line of input contains an integer n, indicating the number of digits in the barcode.

The second line of input contains n space-separated integers, representing the digits of the barcode.

Output format :
The output prints the linked list representing the modified barcode, after adding 1 to its value.



Code constraints :
The barcode represents a non-negative integer.

The number of digits in the barcode can vary.

Sample test cases :
Input 1 :
4
1 9 9 9
Output 1 :
2 0 0 0 
Input 2 :
6
1 2 3 4 5 6
Output 2 :
1 2 3 4 5 7