Create binary_data_linkedlist.cpp
You are developing a smart device that reads binary data from sensors installed in various locations. Each sensor generates a stream of binary data, and the binary values are stored in a linked list. Your task is to implement a function that calculates the decimal value of the binary data represented by the linked list.



The linked list represents the binary representation of a number, with the most significant bit at the head of the linked list. Your function should traverse the linked list and calculate the decimal value of the binary number. 



Note: This is a sample question asked in the Google interview.

Input format :
The input consists of a singly-linked list representing the binary data.

Each node in the linked list contains a value of either 0 or 1.

Output format :
The output is the decimal value calculated from the binary data represented by the linked list.

Code constraints :
The number of nodes will not exceed 30.

Each node's value is either 0 or 1.

The binary data is guaranteed to be a valid representation of a non-negative integer.

The most significant bit is located at the head of the linked list.

Sample test cases :
Input 1 :
3
1 0 1
Output 1 :
Linked List: 1 0 1 
Decimal Value: 5
Input 2 :
0
Output 2 :
Linked List: Empty linked list
Decimal Value: 0