Create Elsa _delete_even_pos_node.cpp
Elsa wants to delete nodes with even positions in a singly linked list. She needs your help to write a program that takes the size of the linked list and its elements as input and uses insertion at the end to add nodes to the linked list and delete nodes with even positions from the linked list. She wants to see the original linked list before deletion and the final linked list after the deletion process.



Write a program to solve Elsa's problem.



Note: This is a sample question asked in a TCS interview.

Input format :
The first line of the input consists of the size (n) of the linked list (an integer).

The second line of the input consists of the elements of the linked list (a sequence of space-separated integers).

Output format :
The output displays the original linked list before the deletion process.

The final linked list after deleting nodes with even positions.



Refer to the sample output for format specifications.

Code constraints :
1<=n<=100

-10000<=arr<=10000

Sample test cases :
Input 1 :
5
100 200 300 10000 9999
Output 1 :
Original Linked List: 100 200 300 10000 9999 
Final Linked List: 100 300 9999 
Input 2 :
3
-10000 -2000 -3000
Output 2 :
Original Linked List: -10000 -2000 -3000 
Final Linked List: -10000 -3000