Create delete_specific_position.cpp
You are developing a program for an online shopping cart system. Each node in the singly linked list represents a product added to the cart. The program should allow for deleting a product from the cart at a specific position and displaying the updated cart.



Write a program to implement the online shopping cart system according to the following specifications:



The program should prompt the user to enter the number of products in the cart.
For each product, the program should prompt the user to enter its product code (an integer value).
The program should create a singly linked list with nodes representing the products in the cart, where each node contains the product code.
The program should then prompt the user to enter the position of the product to be deleted from the cart.
The program should delete the node at the specified position from the linked list.
Finally, the program should traverse the updated linked list and print the product codes of the remaining products in the cart.


Note: This is a sample question asked in a TCS interview.

Input format :
The first line of input consists of an integer n, the number of nodes in the linked list.

The second line of input consists of n integers, representing the data value of each node in the linked list, separated by space.

The third line of input consists of an integer x, representing the position of the node to be deleted. (1-based)

Output format :
The output consists of a single line containing the data values of the modified linked list after deleting the node at position x. The data values should be separated by a space.

Code constraints :
1 ≤ n ≤ 10^5 (number of elements in the linked list)

-10^9 ≤ value of each node ≤ 10^9

1 ≤ x ≤ n (position of the node to be deleted)

Sample test cases :
Input 1 :
7
7270 5318 2964 4237 7388 7010 2468
4
Output 1 :
7270 5318 2964 7388 7010 2468 
Input 2 :
6
10 20 30 40 50 60
3
Output 2 :
10 20 40 50 60