Create contact_management_system.cpp
Problem Statement



You are developing a contact management system for a mobile application. The system allows users to maintain a list of their contacts. Whenever a user adds a new contact, the system should add it at the end of the contact list.



Write a program to implement the code to perform insertion at the end using Grounded Header Linked List.



Note: This is a sample question asked in a TCS interview.

Input format :
The first line represents the size of element n.

The next n lines store the values in it.

Output format :
The output represents the updated list with the nodes added at the end.

Sample test cases :
Input 1 :
5
1
2
3
4
5
Output 1 :
1 2 3 4 5