Create Milton_project.cpp
Milton is working on a project that involves managing data using a circular header-linked list. The circular header linked list is a special type of linked list where the last node points back to the head node. Each node in the list contains a value and a pointer to the next node. 



To ensure the accuracy and reliability of the data, Milton needs to implement a program that checks if the circular header linked list is sorted in non-decreasing order. This program will help Milton identify if any elements are out of order and take appropriate actions to rectify the issue.



Write a program that Milton can use to check if the circular header linked list is sorted in non-decreasing order.



Note: This is a sample question asked in TCS recruitment.

Input format :
The first line represents the number of elements in the linked list n.

The next line represents the n elements in a linked list.

Output format :
The output prints if the circular header linked list is sorted in non-decreasing order or not.

Sample test cases :
Input 1 :
5
1 2 3 4 5
Output 1 :
The list is sorted in non-decreasing order.
Input 2 :
4
1 4 2 3
Output 2 :
The list is not sorted in non-decreasing order.