Create playlist_manager.cpp
You are building a playlist manager for your favorite movies using a circular header-linked list. Each node in the list represents a movie, and the playlist loops back to the beginning after the last movie. 



The playlist manager should allow you to perform operations such as adding movies at the front and end, removing movies from the front and end, and displaying the current playlist.



Note: This is a sample question asked in a TCS interview.

Input format :
The input consists of the following:



The number of movies to insert at the front of the playlist, denoted by an integer n1.

n1 lines containing the titles of the movies to be inserted at the front.

The number of movies to insert at the end of the playlist, denoted by an integer n2.

n2 lines containing the titles of the movies to be inserted at the end.

Output format :
The output consists of the following parts:



The current playlist is displayed as a circular header-linked list.

The number of nodes (movies) in the playlist.

If a movie is deleted from the front or end, the deleted element is displayed.

The updated playlist after deletion.



Refer to the sample output for specifications.

Code constraints :
The movie titles will be alphanumeric strings of a maximum length of 100 characters.

The maximum number of movies in the playlist is 100.

Sample test cases :
Input 1 :
4
Movie 1
Movie 2
Movie 3
Movie 4
2
Movie 5
Movie 6
Output 1 :
|Movie 4| --> |Movie 3| --> |Movie 2| --> |Movie 1| --> |Movie 5| --> |Movie 6| --> 
No of nodes in the CLL is 6
Deleted element: Movie 4
Deleted element: Movie 6
|Movie 3| --> |Movie 2| --> |Movie 1| --> |Movie 5| --> 
No of nodes in the CLL is 4
Input 2 :
3
The Avengers
Iron Man
Thor
0
Output 2 :
|Thor| --> |Iron Man| --> |The Avengers| --> 
No of nodes in the CLL is 3
Deleted element: Thor
Deleted element: The Avengers
|Iron Man| --> 
No of nodes in the CLL is 1
Input 3 :
0
4
Harry Potter and the Prisoner of Azkaban
Harry Potter and the Goblet of Fire
Harry Potter and the Order of the Phoenix
Harry Potter and the Half-Blood Prince
Output 3 :
|Harry Potter and the Prisoner of Azkaban| --> |Harry Potter and the Goblet of Fire| --> |Harry Potter and the Order of the Phoenix| --> |Harry Potter and the Half-Blood Prince| --> 
No of nodes in the CLL is 4
Deleted element: Harry Potter and the Prisoner of Azkaban
Deleted element: Harry Potter and the Half-Blood Prince
|Harry Potter and the Goblet of Fire| --> |Harry Potter and the Order of the Phoenix| --> 
No of nodes in the CLL is 2