Create Lara_linked_list.cpp
Lara is learning about linked lists in her computer science class and wants to practice performing operations on linked lists. Specifically, she's interested in deleting the middle element of a linked list. Can you help her by designing a program that allows her to delete the middle element from a linked list and visualize the result?



Manages a singly linked list.
Implements a function to delete the middle element from the linked list.
Provides a function to push elements onto the linked list.
Displays the elements of the linked list after deleting the middle element.


Note: This is a sample question asked in CTS recruitment.

Input format :
The input consists of an integer n from the user, indicating the number of elements to be pushed onto the linked list.

The linked list is initially filled with integers from n down to 1.

Output format :
The output prints the elements of the linked list after deleting the middle element.



Refer to the sample output for the exact format.

Code constraints :
n should be odd.

3 <= n <= 51

Sample test cases :
Input 1 :
7
Output 1 :
1 2 3 5 6 7 
Input 2 :
11
Output 2 :
1 2 3 4 5 7 8 9 10 11